i18n: redirect 'back' when changing languages
The /lang route was redirecting to '/'. This means that if you
were on a search or callsign page and changed language, you'd
be navigated back to the beginning. A frustrating user experience.

This change uses 'back' instead of '/'. This means that if you
were on a search or callsign page and changed language, you'd
be sent back to the page you were just on and it will be in
a different language. This is what most users like me expect.